+++ src/mesh.rs
@@ -24,7 +24,7 @@ use anyhow::{anyhow, Result};
 /// (see the above link to the axom documentation).
 #[derive(Debug, Clone)]
 pub(crate) struct Mesh {
-    points: Vec<[f32; 2]>,
+    points: Vec<[f64; 2]>,
     cells: Vec<usize>,
     offsets: Offsets,
 }
@@ -65,7 +65,7 @@ impl<'a> Iterator for Cells<'a> {
 impl Mesh {
     /// Constructs a new [`Mesh`] from its array representation.
     pub(crate) fn new(
-        points: Vec<[f32; 2]>,
+        points: Vec<[f64; 2]>,
         cells: Vec<usize>,
         offsets: Vec<usize>,
     ) -> Result<Self> {
@@ -91,7 +91,7 @@ impl Mesh {
     /// The offsets are stored implicitly, which leads to less memory usage, but this is only
     /// possible for single cell type topologies.
     pub(crate) fn with_stride(
-        points: Vec<[f32; 2]>,
+        points: Vec<[f64; 2]>,
         cells: Vec<usize>,
         stride: usize,
     ) -> Result<Self> {
@@ -118,10 +118,10 @@ impl Mesh {
 
     /// Constructs a rectilinear [`Mesh`].
     pub(crate) fn grid(
-        xmin: f32,
-        xmax: f32,
-        ymin: f32,
-        ymax: f32,
+        xmin: f64,
+        xmax: f64,
+        ymin: f64,
+        ymax: f64,
         nx: usize,
         ny: usize,
     ) -> Result<Self> {
@@ -133,13 +133,13 @@ impl Mesh {
             ));
         }
 
-        let dx = (xmax - xmin) / nx as f32;
-        let x: Vec<_> = (0..=nx).map(|i| i as f32 * dx).collect();
+        let dx = (xmax - xmin) / nx as f64;
+        let x: Vec<_> = (0..=nx).map(|i| i as f64 * dx).collect();
         let x = &x.repeat(ny + 1);
 
-        let dy = (ymax - ymin) / ny as f32;
+        let dy = (ymax - ymin) / ny as f64;
         let y: Vec<_> = (0..=ny)
-            .flat_map(|i| [i as f32 * dy].repeat(nx + 1))
+            .flat_map(|i| [i as f64 * dy].repeat(nx + 1))
             .collect();
 
         let points: Vec<_> = x.iter().zip(&y).map(|(&x, &y)| [x, y]).collect();
@@ -199,7 +199,7 @@ impl Mesh {
     }
 
     /// An iterator over the vertices of the [`Mesh`].
-    pub(crate) fn points(&self) -> Iter<[f32; 2]> {
+    pub(crate) fn points(&self) -> Iter<[f64; 2]> {
         self.points.iter()
     }
 }
